Try more stuff

+++ .ci/mingw-64/dist.sh
@@ -3,9 +3,15 @@ make mostlyclean
 deps=`x86_64-w64-mingw32-objdump -p chronicle.exe | grep dll | awk '{print $3}'`
 for dep in ${deps}
 do
-    dep_file=`find /usr -name ${dep} | grep dll -m 1`
-    echo "Copying ${dep_file} to $PWD"
-    cp ${dep_file} .
+    echo "Looking for ${dep}..."
+    dep_file=`find /usr -name ${dep} | grep -m 1 .dll`
+    if [ -e ${dep_file} ] && [ ! -z ${dep_file} ]
+    then
+        echo "Copying ${dep_file} to $PWD"
+        cp ${dep_file} .
+    else
+        echo "Couldn't find it, moving on"
+    fi
 done
 cp ../LICENCE .
 cp ../README.md .
